+++ git-webui/release/share/git-webui/webui/js/git-webui.js
@@ -994,13 +994,13 @@ webui.SideBarView = function(mainView, noEventHandlers) {
                                     '<h4>Discarded Files</h4>' +
                                 '</section>' +
                                 '<section id="sidebar-local-branches">' +
-                                    '<h4 class="mt-3">Local Branches' +
+                                    '<h4 class="mt-1">Local Branches' +
                                     '<button type="button" class="btn btn-default btn-sidebar-icon btn-add shadow-none" >' +
                                         webui.circlePlusIcon+
                                     '</button>' + '</h4>' +
                                 '</section>' +
                                 '<section id="sidebar-remote-branches">' +
-                                    '<h4 class="mt-3">Remote Branches' +
+                                    '<h4 class="mt-1">Remote Branches' +
                                     '<button type="button" class="btn btn-default btn-sidebar-icon btn-prune-remote-branches shadow-none" >'+
                                         webui.refreshIcon+
                                       '</button>' +'</h4>' +
@@ -1008,6 +1008,7 @@ webui.SideBarView = function(mainView, noEventHandlers) {
                                 '<section id="sidebar-tags">' +
                                     '<h4>Tags</h4>' +
                                 '</section>' +
+                                '<section id="space-filler"></section>'+
                                 '<section id="sidebar-settings">' +
                                     '<h4>Settings</h4>' +
                                 '</section>' +
@@ -2600,13 +2601,20 @@ webui.NewChangedFilesView = function(workspaceView) {
                 });
                 $("#commitBtn").off("click");
                 $("#commitBtn").on("click", function() {
-                    if (selectedItemsFromOtherUser.length > 0) {
-                        self.confirmActionOnOtherUsersChanges("commit");
-                    } else {
-                        var commitMessage = $('#commitMsg').val();
-                        self.commit(commitMessage, $("#commitMsgDetail").val());
-                    }
-                    
+                    // Make sure we are not commiting to default merge branch in basic mode
+                    $.get("api/basic-and-default", function (data) {
+                        var basicAndDefault = JSON.parse(data)["basic-and-default"]
+                        if (basicAndDefault == "1") {
+                            self.noCommitsOnDefault();
+                        } else {
+                            if (selectedItemsFromOtherUser.length > 0) {
+                                self.confirmActionOnOtherUsersChanges("commit");
+                            } else {
+                                var commitMessage = $('#commitMsg').val();
+                                self.commit(commitMessage, $("#commitMsgDetail").val());
+                            }
+                        }
+                    })                    
                 });
 
                 $("#amendBtn").off("click");
@@ -2750,6 +2758,55 @@ webui.NewChangedFilesView = function(workspaceView) {
             removePopup(popup);
         });
     }
+
+    // Popup for when trying to commit to default merge branch in basic mode
+    self.noCommitsOnDefault = function () {
+        function removePopup(popup) {
+            $(popup).children(".modal-fade").modal("hide");
+            $(".modal-backdrop").remove();
+            $("#noCommitsDefault").remove();
+        }
+
+        var popup = $(
+            '<div class="modal fade" tabindex="-1" id="noCommitsDefault" role="dialog" data-backdrop="static">' +
+                '<div class="modal-dialog modal-md" role="document">' +
+                    '<div class="modal-content">' + 
+                        '<div class="modal-header">' +
+                            '<h5 class="modal-title">Cannot commit to Default Branch</h5>' +
+                            '<button type="button" class="btn btn-default close" data-dismiss="modal">' + webui.largeXIcon + '</button>' +
+                        '</div>' +
+                        '<div class="modal-body">' + 
+                            '<div class="row">' +
+                                '<div class="col-sm-1">' +
+                                    webui.warningIcon +
+                                '</div>' +
+                                '<div class="col-sm-11">' +
+                                    '<p>You cannot commit directly to the default merge branch while using basic mode. Please switch to another branch.</p>' + 
+                                '</div>' +
+                            '</div>' +
+                        '</div>' +
+                        '<div class="modal-footer"></div>' +
+                    '</div>' + 
+                '</div>' +
+            '</div>'
+        )[0];
+
+        $("body").append(popup);
+
+        var popupFooter = $(".modal-footer", popup)[0];
+        webui.detachChildren(popupFooter);
+
+        $(
+            '<button class="btn btn-sm btn-secondary action-btn" id="noCommitDefaultButton">Ok</button>'
+        ).appendTo(popupFooter);
+
+        $(popup).modal('show');
+
+        $("#noCommitsDefault").find(".close, #noCommitDefaultButton").click(function() {
+            removePopup(popup);
+        })
+
+    };
 
     self.confirmActionOnOtherUsersChanges = function(action) {
             function removeWarningModal(popup) {