+++ CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed display of other users' username in workspace view on Unix (#530)
+- Fix left-sidebar spacing (#525)
 - Fixed slowness loading some CSP pages with multiple instances sharing a webserver (#540)
+- Prevent direct commits to default merge branch in basic mode (#484)
 - Fixed GetContexts utils function to exclude implied namespaces from the list of namespaces(#468)
 
 ## [2.6.0] - 2024-10-07